Move plugin-specific opi target views to plugin packages
