+++ base/uk.ac.stfc.isis.ibex.ui.synoptic/plugin.xml
@@ -22,8 +22,8 @@
       </view>
       <view
             allowMultiple="true"
-            class="uk.ac.stfc.isis.ibex.ui.targets.SynopticOpiTargetView"
-            id="uk.ac.stfc.isis.ibex.ui.targets.SynopticOpiTargetView"
+            class="uk.ac.stfc.isis.ibex.ui.synoptic.views.SynopticOpiTargetView"
+            id="uk.ac.stfc.isis.ibex.ui.synoptic.views.SynopticOpiTargetView"
             name="Component view"
             restorable="true">
       </view>
@@ -72,7 +72,7 @@
          </view>
          <view
                closeable="true"
-               id="uk.ac.stfc.isis.ibex.ui.targets.SynopticOpiTargetView"
+               id="uk.ac.stfc.isis.ibex.ui.synoptic.views.SynopticOpiTargetView"
                minimized="false"
                moveable="false"
                ratio="0.6f"
