+++ base/uk.ac.stfc.isis.ibex.ui.synoptic/src/uk/ac/stfc/isis/ibex/ui/synoptic/SynopticPresenter.java
@@ -53,8 +53,8 @@ import uk.ac.stfc.isis.ibex.targets.Target;
 import uk.ac.stfc.isis.ibex.targets.ViewTarget;
 import uk.ac.stfc.isis.ibex.ui.UI;
 import uk.ac.stfc.isis.ibex.ui.synoptic.views.LinkedViews;
+import uk.ac.stfc.isis.ibex.ui.synoptic.views.SynopticOpiTargetView;
 import uk.ac.stfc.isis.ibex.ui.synoptic.views.SynopticView;
-import uk.ac.stfc.isis.ibex.ui.targets.SynopticOpiTargetView;
 
 /**
  * Responsible for the presentation logic of the synoptic.
